Move distributeSurplus to .Balance module
